Fix hashkey error when restarting server
In the case where the git data was not modified since last server
start we didn't set the server hash key to the proper value.
This in turn made all queries fail.
I didn't notice this issue before probably because the disk cache was
returning content before the query engine was even called.
This time I cleared all cache files so the issue was visible.